+++ apps/docs/stories/display.stories.tsx
@@ -16,7 +16,7 @@ import {
   Skeleton,
   Small,
   Spinner
-} from "@js-solutions/react-ui";
+} from "@zetahdev/react-ui";
 
 const meta = {
   title: "Foundations/Display",