+++ packages/react-ui/src/components/checkbox.tsx
@@ -1,5 +1,5 @@
 import * as CheckboxPrimitive from "@radix-ui/react-checkbox";
-import { CheckIcon } from "@js-solutions/icons";
+import { CheckIcon } from "@zetahdev/icons";
 import { forwardRef, type ComponentPropsWithoutRef, type ElementRef } from "react";
 
 import { cn } from "../utils";