+++ packages/react-ui/src/components/dropdown-menu.tsx
@@ -1,5 +1,5 @@
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu";
-import { CheckIcon } from "@js-solutions/icons";
+import { CheckIcon } from "@zetahdev/icons";
 import { forwardRef, type ComponentPropsWithoutRef, type ElementRef } from "react";
 
 import { cn } from "../utils";