+++ packages/react-ui/src/components/menubar.tsx
@@ -1,5 +1,5 @@
 import * as MenubarPrimitive from "@radix-ui/react-menubar";
-import { CheckIcon } from "@js-solutions/icons";
+import { CheckIcon } from "@zetahdev/icons";
 import { forwardRef, type ComponentPropsWithoutRef, type ElementRef } from "react";
 
 import { cn } from "../utils";